Trait pallet_messages_rpc::MessagesApiServer

pub trait MessagesApiServer<BlockHash, BlockNumber>: Sized + Send + Sync + 'static {
    fn get_messages_by_schema(
        &self,
        schema_id: SchemaId,
        pagination: BlockPaginationRequest<BlockNumber>
    ) -> RpcResult<BlockPaginationResponse<BlockNumber, MessageResponse<BlockNumber>>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static,
        BlockNumber: Send + Sync + 'static + DeserializeOwned + Serialize,
    { ... }
}

Server trait implementation for the MessagesApi RPC API.

Required Methods

fn get_messages_by_schema(
&self,
schema_id: SchemaId,
pagination: BlockPaginationRequest<BlockNumber>
) -> RpcResult<BlockPaginationResponse<BlockNumber, MessageResponse<BlockNumber>>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self> where
BlockHash: Send + Sync + 'static,
BlockNumber: Send + Sync + 'static + DeserializeOwned + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.