Trait pallet_messages_rpc::MessagesApiServer

pub trait MessagesApiServer: Sized + Send + Sync + 'static {
    fn get_messages_by_schema_id(
        &self,
        schema_id: SchemaId,
        pagination: BlockPaginationRequest
    ) -> RpcResult<BlockPaginationResponse<MessageResponse>>;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the MessagesApi RPC API.

Required Methods

fn get_messages_by_schema_id(
&self,
schema_id: SchemaId,
pagination: BlockPaginationRequest
) -> RpcResult<BlockPaginationResponse<MessageResponse>>

Retrieve paginated messages by schema id

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.